changes plot updates mechanism to use win-plot-refs
Now the plot does not store a window ID any more. When an update
of the plot is needed a reverse lookup is made in the window-plot
refs table to find all the windows that references the given plot
and asking the to update the slot.

The net results of this change is that:
- all the windows that contains the plot are automatically updated
- the lua_plot class is now virtually a simple "plot" class
- the lookup mechanisms could be slower because we need to search
  the windows by inspecting some lua tables

Added also the missing 'show' method for plots.