new model for C++/Lua error exception-safe separation
Now all the exceptions are trapped in pure C++ functions and a status
code is returned by reference to indicate success or give an error
message. The C function that instruments Lua interface can call lua_error
but is not supposed to catch exceptions and does not allocate auto class
instance that need non-trivial destructors.