changed window closing unref main lock strategy
A global lock is used now to guard the main loop. The plot windows directly
unrefs the window by just waiting the main lock. Simplified also the
window index code since plot are not indexed since many versions now.

The windows are not explicitely closed when the main loop terminates.