Introduced window registry and changed win/plot weak table to ordinary.
Now the table that store the window <-> plot relations is no more weak on
the keys (windows). The windows are removed from the table when the event
loop thread of the windows terminates. In this way we don't depend on the
GC and we avoid the problems with delayed removal of the keys.