Changed default multiplication operator to perform matrix multiplication.
The new semantic associate the multiplication operator to the matrix
multiplication. The element-wise multiplication is no more supported. The
'mul' function disappears because it is redundant with multiplication
operator. In order to support all the type of multiplication between
real/complex numbers and real/complex matrix the arithmetic operators of
numbers and matrix is now unified.