Eliminate C++ exceptions from graphical system implementation
The rationale is that everything is simpler without C++ exceptions.
No exceptions are raised in the C++ code and errors are handler in the
tradition C fashion. lua_error is not directly called from C++ functions
that may allocate a non POD object on the stack.

If the C++ "new" operator raise an exception the program will abort. The
idea is that handle such an improbable occurrence greatly complicates the
code. Where big chunks of memory are allocated the nothrow variant of
"new" is used, the returned pointer is verified and an error is raised if
the allocation failed.

In the Windows code the bitmap allocation for image may raise an unhandled
exception in there is not enough memory. The memory allocation is done
inside the AGG library so it difficult to change its behavior.