Improve the procedure to properly close the application
Now when the console window is closed the exit signal is sent to the Lua
thread but the window and the FXApp continue to run. When the Lua thread
actually terminate the current Lua command it does send a signal to the
FOX thread and call pthread_exit without closing the Lua state. The Lua
state is then closed by the main FOX thread when the closing signal from
Lua is handled.