Emit UCLO instructions when local variables goes out of scope
The UCLO instruction should be emitted when a block is left if there
was local variables declared in the block. This is required because a
closure may refer to the block specific local variables and since these are
going out of scope they should be closed when the block is left.

The flag need_close is also not cleared when a return instruction is
treated because other branches of the code may need to emit UCLO
instructions as well.