Modify gsl shell's REPL to use the new bytecode compiler
The new bytecode compiler is based on those of the Nyanga language. The
pipeline include the lexer, the parser and the bytecode generator all
implemented in pure Lua.

Some C functions are used to start a dedicated Lua state for the
parser/compiler. Then the REPL use the functions loadbuffer and loadfile
to feed data to the Lua parser and obtain the compiled bytecode.