Implement language function to pass a list of "use" modules
Feature to be used to implement the interactive "use" list for the REPL.

A linked list of module names can be passed to language_loadbuffer_use_ext.
In turns the function pass the modules list to the parser that issues
appropriate "use" statements in the main "chunk" built by the parser.