Adopt the character ':' to identify the goto labels
The character ':' is therefore used both for slicing and for goto
labels. There are no conflict at the level of the parser because goto
labels are statements while matrix slicing can occur only in
expressions' indexing.

The symbol for the objects' methods call remain '::'. To identify
this latter the token name 'TK_resolve' is introduced into the lexer.

For the sake of code clarity the symbols RANGE_TOKEN, LABEL_TOKEN,
RESOLVE_TOKEN are introduced in the parser to make clear what parsing
rule applies. Actually RANGE_TOKEN and LABEL_TOKEN correspond to the
same token ':'.