Implement matrix slicling with implicit start or ending indexes
This will allow to write m[:,:] or m[:n, a:]. The semantic if the same
of Python, if the initial index is not given it is implicitly taken
to 1. If the last index is got given the slicing is made up to the
matrix's size.

When the incomplete slicing is used the generated AST is always
completed and a literal one, m.size1 or m.size2 is placed in the AST
for the missing terms.