Implement support for keyword arguments function calls
The keyword arguments functions are functional, you can call a
function with the syntax "foo(x, something = 3)" and it will be
translated to foo.__keyargs({something = 3}, x).

When no keyword arguments are given the function call is performed
normally, without any translations.

The function "expr_keyword" introduced into the parser is similar to
what used to parse table literals and may be we can better reuse the
code. The problem is that the table literals accept the form
[<expr>] = <something> which is not accepted for keyword arguments
calls.

The code for expr_keyword does not take into account the LJ5.2
gymnastic about the "goto" keyword.