Introduce a check to verify that a value is numeric in var_declare
Introduce also the function var_declare_num to declare a numeric
variable without a specific value to be used only in numeric "for"
loop.

With this modification the "num_const" field will be "true" only if
a variable is non-mutable and is known to be numeric.

NB: There is a problem with for loops because an outer local variable
can be changed in the body of the for loop *after* the check and yet
the value should be considered mutable because the loop came back
to the initial instruction. This problem is not related to this
commit.