Evaluate keyword arguments defaults at function's declaration
The defaults value are evaluated only once when the keyword arguments
function declaration is encountered.

This is more logical because from the lexical point of view the
programmer will naturally expect that the defaults value to be
evaluated only once when the declaration occour.

The current schema excludes the utilisation of keyword arguments for
function as expressions.