Remove the need for graphics windows to access Lua state on closure
Now the closed windows are removed from the windows' list with a function
called when opening a new window.

The windows no longer needs to access the Lua state and so the global gsl
shell state in removed simplifying the architecture of the module.

In addition the lock needed to manage the shutdown of all the windows is
no longer needed since now the windows does not attempt to access the
Lua state when closing.