Separate window and plotting libraries
The functions to create window are now in a separate shared library
from the plotting functions.

The plotting module is loaded by Lua and a stub windows library is
initially loaded. Then, if the "native window" module is loaded the
plotting module and the windows module are tied together.

The links is done using a light userdata pointer to a C struct
containing a couple of pointer to C functions. The struct is hold
in the registy using the key "github.com/franko/libgraph".