Remove all agg-plot stuff
Remove all agg-plot based graphics stuff.

The AGG library remained as a dependency because if used in FOX app for
non-graphical stuff but can be removed easily.

It compiles and run properly but the plotting capability related to the
gdt module no longer works. The code is still there yet but not loaded.