Implements a preliminary version of boxplot
The boxplot are drawn incorrectly, they doesn't follow the correct rule
and the outliers are not drawn.

The implementation needs some cleanup to avoid come code duplication.

A problem is inherited from the barplot implementation: the x positions
of the bars is not correct versus the x axis labels when multiple factors
are cumulated in the x axis.