Add GSL library as a subproject
The goal is to link the GSL library using the --link-whole option and
use the --export-symbol on the executable so that all GSL functions
are available to the application using the LuaJIT's FFI extension.

Currently Meson seems to support the as_link_whole() as an option
only for dependency from declare_dependency so GSL needs to be a
subproject.