Add functions activate and chdir in project module
The function "activate" in the newly introduced module "project" has the
purpose to let a script access project's modules and data without
hard-coding their paths.

When "activate" is used the application change the working directory to
the project's directory and add the current directory into the
package.path variable.

Expose also the "chdir" function because it may be useful for other
purposes.

The implementation works on Linux but needs to be tested on Windows.