Implement 2D interpolations module GSL functions
It works correctly and the interface is fine.

Some functions declared from the GSL headers are not actually used and
may be removed.

The module could be unified with the 1D interpolation module. This
latter does not exist yet but 1D interpolation is done from the GDT
module, for tables.