Implement intermediate values logger for ODE
Introduce the record module to store time-dependent values and plot the
stored values.

The ODE solver now use the record module to let the ODE function store
any kind of intermediate values using a call:

log:store(t, {value1 = xx, value2 = xx})

The "log" argument will be provided to the function as the last
argument. If it is nil no values should be stored so the function should
always check if it is nil before storing values.

The log will be actually passed when the ODE solver crosses the t
intermediate values required by the "evolve" function.