Move cursor into input line on key press
To improve usability, when the user press a key without a control
modifier, the cursor is moved to the input line so that the text is
entered there.

In addition, as a special rule, is the paste command is used from a
selection above the input line, the text is pasted at the end of the
input line.

Related to github issue:

https://github.com/franko/gsl-shell/issues/36